Renaming immunisation -> vaccine_course
