+++ server/graphql/vaccine_course/src/lib.rs
