+++ server/graphql/vaccine_course/src/types/mod.rs
