+++ server/repository/src/db_diesel/vaccine_course/vaccine_course_item.rs
