+++ server/repository/src/db_diesel/vaccine_course/vaccine_course_row.rs
