+++ server/repository/src/db_diesel/vaccine_course/vaccine_course_schedule_row.rs
